move zipfFileIO into own header route loadTextureFile from fileIO plugin fix B3_ENABLE_FILEIO_PLUGIN logic
