+++ examples/SharedMemory/plugins/tinyRendererPlugin/TinyRendererVisualShapeConverter.h
@@ -47,7 +47,7 @@ struct TinyRendererVisualShapeConverter : public UrdfRenderingInterface
 	virtual void render();
 	virtual void render(const float viewMat[16], const float projMat[16]);
 
-	virtual int loadTextureFile(const char* filename);
+	virtual int loadTextureFile(const char* filename, struct CommonFileIOInterface* fileIO);
 	virtual int registerTexture(unsigned char* texels, int width, int height);
 
 	virtual void syncTransform(int shapeUid, const class btTransform& worldTransform, const class btVector3& localScaling);
